# Optimize Base.valid16?/2